requirement.txt
streamlit 
pypdf
einops
python-docx
google
google-genai
langchain
langchain-community
sentence-transformers